Added wallet address as querystring and renamed mnemonic to a friendly name.
