+++ lib/wallet_create_page.dart
@@ -1,4 +1,4 @@
-import 'package:etherwallet/components/copyButton/copyButton.dart';
+import 'package:etherwallet/components/copyButton/copy_button.dart';
 import 'package:etherwallet/components/form/paper_form.dart';
 import 'package:etherwallet/components/form/paper_input.dart';
 import 'package:etherwallet/components/form/paper_validation_summary.dart';
@@ -47,7 +47,7 @@ class _WalletCreatePage extends State<WalletCreatePage> {
           child: Column(
             children: <Widget>[
               Text(
-                "Get a piece of papper, write down your mnemonic and keep it safe. This is the only way to recover your funds.",
+                "Get a piece of papper, write down your seed phrase and keep it safe. This is the only way to recover your funds.",
                 textAlign: TextAlign.center,
               ),
               Container(
@@ -109,8 +109,8 @@ class _WalletCreatePage extends State<WalletCreatePage> {
             children: <Widget>[
               PaperValidationSummary(widget.store.errors),
               PaperInput(
-                labelText: 'Confirm mnemonic',
-                hintText: 'Please type mnemonic again',
+                labelText: 'Confirm your seed',
+                hintText: 'Please type your seed phrase again',
                 maxLines: 2,
                 onChanged: widget.store.setMnemonicConfirmation,
               ),
