--- todo.md
@@ -1,12 +1,0 @@
-
-[ ] Add support to IOS. (add google-services file and camera permissions)
-[ ] Handle when connection is not available.
-[ ] Better solution to MOBX and form field
-[X] Remove contract abi from dart and use the original json file.\
-[ ] Change configuration per environment (debug/release)\
-[ ] TODO: Contract transactions list\
-[ ] Export account private key / display mnemonic\
-[X] Use flutter name convention for files, which is lowercase_with_underscores.\
-[ ] Change sendTransaction implementation when possible, receiving a stream with confirmations and receipt straight from send method.
-[ ] Add support to add use multiple accounts\
-[ ] c:\flutter\.pub-cache\hosted\pub.dartlang.org\firebase_analytics-4.0.2\android\src\main\java\io\flutter\plugins\firebaseanalytics\FirebaseAnalyticsPlugin.java uses or overrides a deprecated API
